Move Activity.Current into main file
